refactor: rename demon namespace to daemon
