+++ tests/NetDaemon.Daemon.Tests/Daemon/ExtensionMethodsTests.cs
