+++ tests/NetDaemon.Daemon.Tests/Daemon/HttpHandlerMock.cs
