+++ tests/NetDaemon.Daemon.Tests/Daemon/HttpTests.cs
